Разнесла модели по разным файлам
